fix(ai): cast nanoTime longs to double before subtraction in MobSystem stats
Avoids potential overflow when frameTimeStat consumes the delta as a
double-typed metric.

+++ infinity/src/main/java/infinity/ai/MobSystem.java
@@ -188,7 +188,7 @@ public class MobSystem extends AbstractGameSystem {
             long start = System.nanoTime();
             scheduler.update(time);
             long end = System.nanoTime();
-            frameTimeStat.updateValue(end - start);
+            frameTimeStat.updateValue((double) end - (double) start);
             activeMobCountStat.updateValue(brains.size());
         } else {
             scheduler.update(time);
